<camipro> Updated bean
git-svn-id: file:///var/pocketcampus/trunk@776 fd477b11-563a-408f-9d6e-eeb9ca009d0c

+++ shared/src/org/pocketcampus/shared/plugin/camipro/TransactionBean.java
@@ -11,9 +11,9 @@ public class TransactionBean {
 	private String type;
 	private String description;
 	private Date date;
-	private float amount;
+	private double amount;
 	
-	public TransactionBean(String type, String description, Date date, float amount) {
+	public TransactionBean(String type, String description, Date date, double amount) {
 		this.type = type;
 		this.description = description;
 		this.date = date;
@@ -38,7 +38,7 @@ public class TransactionBean {
 	public void setDate(Date date) {
 		this.date = date;
 	}
-	public float getAmount() {
+	public double getAmount() {
 		return amount;
 	}
 	public void setAmount(float amount) {
